+++ app/Models/Admin/Api/Nhanh.php
@@ -1,0 +1,34 @@
+<?php
+
+namespace App\Models\Admin\Api;
+
+use Illuminate\Support\Facades\Log;
+use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\Relations\HasMany;
+use Illuminate\Database\Eloquent\Relations\BelongsTo;
+use Illuminate\Database\Eloquent\Factories\HasFactory;
+use App\Models\Sys\Env\Domain;
+
+class Nhanh extends MoApp\Models\Admin\Api
+{
+    /**
+     * Make attributes available in the json response.
+     *
+     * @var array
+     */
+    protected $appends = [];
+
+    /**
+     * The model's default values for attributes.
+     *
+     * @var array
+     */
+    protected $attributes = [];
+
+    /**
+     * Indicates if the model should be timestamped.
+     *
+     * @var bool
+     */
+    public $timestamps = true;
+}