+++ database/migrations/2023_04_11_132037_create_apis_table.php
@@ -11,12 +11,14 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('vms', function (Blueprint $table) {
+        Schema::create('apis', function (Blueprint $table) {
             $table->id();
-            $table->string('vm_id')->unique();
-            $table->string('name')->nullable();
-            $table->string('pri_host')->nullable();
-            $table->string('pub_host')->nullable();
+            $table->string('type');
+            $table->string('accessToken');
+            $table->string('expiredDateTime')->nullable(); // Y-m-d H:i:s
+            $table->integer('businessId')->nullable();
+            $table->string('depotIds')->nullable();
+            $table->text('permissions')->nullable();
             $table->timestamps();
         });
     }
@@ -26,6 +28,6 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('vms');
+        Schema::dropIfExists('apis');
     }
 };