+++ resources/views/admin/vm/edit.blade.php
@@ -1,0 +1,4 @@
+@extends('admin.layout')
+@section('main')
+    <h3>Api Config</h3>
+@endsection