+++ routes/web.php
@@ -27,6 +27,7 @@ Route::domain("admin.diepxuan.com")->group(function () {
         Route::get('/', [App\Http\Controllers\Admin\HomeController::class, "index"]);
 
         Route::get('/etc/{env}/{vm_id}', [App\Http\Controllers\Sys\EnvController::class, "showByVm"]);
+        Route::get('/api/{type}', [App\Http\Controllers\Admin\ApiController::class, "token"]);
 
         Route::resources([
             'etc' => App\Http\Controllers\Sys\EnvController::class,
@@ -35,6 +36,7 @@ Route::domain("admin.diepxuan.com")->group(function () {
 
         Route::namespace('App\Http\Controllers\Admin')->prefix('admin')->name('admin.')->group(function () {
             Route::resource('vm', VmController::class);
+            Route::resource('api', ApiController::class);
         });
     });
 });